make corpse an encounter (executable program)
Created a hidden file in main directory (.functions) for reusable functions. This is
included in all monsters now to replace the gameover function.
The gameover function now generates a uniq corps name and makes it executable. Once
the corpse is looted, a hidden file is create to prevent looting multiple times.